당첨 여부를 확인하는 PrizeCounter 클래스 생성
- LottoGame 에서 당첨 결과를 반환해주는 책임이 있어 당첨 여부를 확인하는 책임을 PrizeCounter 를 생성하여 위임해주었습니다.
